% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ezmodel.R
\name{regularization_plot}
\alias{regularization_plot}
\title{Plots coefficients from regularizing various models.}
\usage{
regularization_plot(model, lambda, tol = 1e-07, x, y)
}
\arguments{
\item{model:}{String specifying lasso, ridge regression, or logistic regression with L2-regularization.
Argument should be one of "lasso", "ridge", or "logistic".}

\item{lambda:}{Vector of penalty constant(s) multiplying the regularization term. Larger value corresponds to stronger regularization.}

\item{tol:}{Coefficients less than this will be treated as zero.}

\item{x:}{n x d dataframe of features.}

\item{y:}{n x 1 dataframe of response values.}
}
\value{
ggplot object. Plot returned depends on length of lambda.
    length(lambda)==1: Plot displays magnitude of model coefficients, where coefficients with magnitude less than
                       `tol` are treated as zero.
    length(lambda)>1: Plot displays counts of nonzero coefficients in each model, where coefficients with magnitude
                      less than `tol` are treated as zero.
}
\description{
Plots coefficients from regularizing various models.
}
