% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ezmodel.R
\name{score}
\alias{score}
\title{Used to compute an arbitrary score method on arbitrary inputs.}
\usage{
score(model, score_type, train_settings = trainControl(method = "none"))
}
\arguments{
\item{model}{(char): Model type that can be passed into caret's `train()` function.}

\item{score_type}{(char): String specifying score method to be used. Should be one of (mse, accuracy, r2, adj_r2, auc, ...).}

\item{train_settings}{(list): default=trainControl(method='none'). trainControl object containing any desired settings for caret's train() function.}
}
\value{
Function that takes dataframes X (features, n x d) and Y (response, n x 1) that score will be computed on.
}
\description{
Used to compute an arbitrary score method on arbitrary inputs.
}
\examples{
rf_sens_func <- score('rf','sensitivity')
rf_sens_val <- score('rf', 'sensitivity')(x, y)
}
