% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ezmodel.R
\name{train_test_plot}
\alias{train_test_plot}
\title{Creates plot of training and test error for trained model.}
\usage{
train_test_plot(model, score_type, x, y, hyperparameter, param_range,
  random_seed)
}
\arguments{
\item{model:}{Currently only works with argument 'decision_tree'.
Generally:
String specifying decision_tree, lasso, ridge regression, or logistic regression.
Argument should be one of "decision_tree", "lasso", "ridge", or "logistic"..}

\item{score_type:}{(list or str): Should be one of (mse, r2, adj_r2, auc, ...).
If a vector, then a vector containing several of those entries as elements}

\item{x:}{n x d dataframe containing features}

\item{y:}{n x 1 dataframe containing response values.}

\item{hyperparameter:}{string defining hyperparameter to iterate over}

\item{param_range:}{vector of hyperparameter values to iterate over}

\item{random_seed:}{Default = None. If set to integer, defines the random train_test_split}
}
\value{
ggplot object showing training and test score vs. hyperparameter values.
}
\description{
Creates plot of training and test error for trained model.
}
\examples{
train_test_plot(model = "decision_tree", score_type = "accuracy", x = Y,

}
